+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-04-04 20:28:27
-// Commit: feat(test): add bi-directional event bus handler for test
+// Updated at: 2026-04-04 22:39:49
+// Commit: perf(test): optimize memory consumption in test list viewport
 
 class testComponentState {
-  final String commitRef = '127';
-  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
-  final DateTime timestamp = DateTime.parse('2026-04-04T20:28:27+05:00');
+  final String commitRef = '135';
+  final String lastMessage = 'perf(test): optimize memory consumption in test list viewport';
+  final DateTime timestamp = DateTime.parse('2026-04-04T22:39:49+05:00');
 }